v0.2.2 auto generate clientid if none is specified
If the `:client` option is nil, or unspecified, when starting a client
via `start/3` or `start_link/3` a randomly generated one will get used
instead.

This is good if you want to connect multiple times to the same broker
and don't want/need to specify the client id for each of the
connections.

If a client id is set it will get respected, so it is still possible to
generate them to your applications needs.